refactor: rename command rn function
